Create README.md
In this Case study, we're aiming for a real business scenario. Applying Exploratory Data Analysis (EDA) will give out the ability to understand and statistically interpret data to derive actionable insight from the bank.

The main aims of this study will be: to identify patterns that indicate if a client has difficulty paying their installments which may be used for taking actions such as denying the loan, reducing the amount of loan, and lending (to risky applicants) at a higher interest rate, etc. This will ensure that the consumers capable of repaying the loan are not rejected. Identification of such applicants using EDA is the aim of this case study.

I recommend my data analyzer audience to run this command in google collaborator. Good luck enjoys the project...